requirements.txt
Django==1.10.5
django-guardian
djangorestframework==3.5.4
markdown==2.6.8
django-filter==1.0.1
psycopg2
pycrypto
djangorestframework-jwt